Add Convex Webhooks support from video series
This will need to be run, with env variables for the Convex project, in
order to generate the requisite APIs.

```
npx convex dev
```

Additionally, the `npx convex dev` will need to be added to a pre-build
script, so `convex/_generated` is guaranteed to be in place.

Publically exposed endpoints.
Exposes `/workos-webhook`.

Convex schema.
Currently only defines `email` and `workos_id`.

Defines application SDK for managing users in app code.
These APIs are exposed to the client and can therefore be manipulated.

Defines "internal" APIs that can only be called from other Convex
Actions.
It *must* be in a separet file to utilize the `"use node"` directive.
Otherwise it will fail on `crypto` import.